+++ backend/db/sqlc/score_interface.go
@@ -8,11 +8,11 @@ import (
 type ScoreInterface interface {
 	GetScoreID() string
 	GetUserID() string
-	GetStage() int32
+	GetStage() int8
 	GetPairname() string
 	GetEntrytime() string
 	GetPosition() string
-	GetLeverage() int32
+	GetLeverage() int8
 	GetOuttime() int64
 	GetEntryprice() float64
 	GetQuantity() float64
@@ -25,11 +25,11 @@ type ScoreInterface interface {
 
 func (s *PracScore) GetScoreID() string         { return s.ScoreID }
 func (s *PracScore) GetUserID() string          { return s.UserID }
-func (s *PracScore) GetStage() int32            { return s.Stage }
+func (s *PracScore) GetStage() int8             { return s.Stage }
 func (s *PracScore) GetPairname() string        { return s.Pairname }
 func (s *PracScore) GetEntrytime() string       { return s.Entrytime }
 func (s *PracScore) GetPosition() string        { return s.Position }
-func (s *PracScore) GetLeverage() int32         { return s.Leverage }
+func (s *PracScore) GetLeverage() int8          { return s.Leverage }
 func (s *PracScore) GetOuttime() int64          { return s.Outtime }
 func (s *PracScore) GetEntryprice() float64     { return s.Entryprice }
 func (s *PracScore) GetQuantity() float64       { return s.Quantity }
@@ -41,11 +41,11 @@ func (s *PracScore) GetCreatedAt() time.Time    { return s.CreatedAt }
 
 func (s *CompScore) GetScoreID() string         { return s.ScoreID }
 func (s *CompScore) GetUserID() string          { return s.UserID }
-func (s *CompScore) GetStage() int32            { return s.Stage }
+func (s *CompScore) GetStage() int8             { return s.Stage }
 func (s *CompScore) GetPairname() string        { return s.Pairname }
 func (s *CompScore) GetEntrytime() string       { return s.Entrytime }
 func (s *CompScore) GetPosition() string        { return s.Position }
-func (s *CompScore) GetLeverage() int32         { return s.Leverage }
+func (s *CompScore) GetLeverage() int8          { return s.Leverage }
 func (s *CompScore) GetOuttime() int64          { return s.Outtime }
 func (s *CompScore) GetEntryprice() float64     { return s.Entryprice }
 func (s *CompScore) GetQuantity() float64       { return s.Quantity }
